+++ src/tts/gather-text.ts
@@ -411,6 +411,13 @@ export class TextGather {
             },
         })
 
+        sc.TutorialMarkerGui.inject({
+            init(x, y, width, height, text, direction, connectPos, stopTime, doneCallback) {
+                this.parent(x, y, width, height, text, direction, connectPos, stopTime, doneCallback)
+                speakIC(text)
+            },
+        })
+
         /* ------------------ menu ------------------ */
         let ignoreButtonFrom: number = 0