+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # Change Log
 
 ## [Unreleased]
+
 ## [0.5.6] 2024-02-04
 
 ### Added
@@ -10,6 +11,7 @@
 - Announce walking onto hints
 - Added recived item announcing
 - Add "Floating Moving Platform" puzzle entity hint
+- Added info sign hint
 
 ### Fixed